refactor(parser): split math-extractor tests by knowledge domain
Split the 554-line test file into two focused modules:
- math-extractor.browser.ts (348 lines): renderer-specific extraction
  tests (KaTeX, MathJax v3, MathJax v2, Native MathML)
- math-extractor-edge-cases.browser.ts (212 lines): cross-format edge
  cases and preprocessing pipeline tests
